Use cadvisor docker API and not container API
This is to address container stats not working on RancherOS.  The real
issue is that Docker-in-Docker was not working as a result of the fact
that we were assuming a non-nested Docker cgroup path.  On Docker 1.7
the paths were always /docker/id, now they are /docker/id/docker/id if
nested.